requirements.txt
discord.py==2.3.2
google-generativeai>=0.7.0
aiohttp>=3.9.0
